# docs/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: TransEV EV CMS API
  version: 0.11.0
  description: |
    Canonical machine-readable contract for the endpoints currently implemented
    by `ev-cms-backend-new`.

    Authentication uses a signed-then-encrypted bearer access token. Platform
    sessions and CPO sessions are separate. CPO integration operations also
    require `X-CPO-App-ID`; that header is checked only after the bearer token
    establishes the CPO and never selects a tenant.

    Public customer signup requires `X-CPO-App-ID` to resolve an active CPO, but
    the header remains public routing metadata rather than authentication.
    Administrative authentication, platform control-plane, health, and
    documentation routes do not require it. Request bodies reject unknown
    properties and are limited to 32 KiB by the server.

    Every request that reaches Gin returns a server-generated UUID in the
    `X-Request-ID` response header for correlation with the safe structured
    completion log. The logger records matched route templates and stable
    handled error codes, never bodies, raw paths, queries, credentials, tokens,
    OTPs, email addresses, user agents, app IDs, API messages, or panic values.
    `LOG_LEVEL=DEBUG` adds request-start and handled-error component/type
    diagnostics under the same request ID without relaxing those exclusions.
  contact:
    name: TransEV Engineering
servers:
  - url: /
    description: Same origin as the API documentation page
...
paths:
...
  /api/v1/app/auth/signup:
    post:
      tags: [App User Authentication]
      operationId: startCustomerSignup
      summary: Start CPO-scoped customer signup
      description: |
        Resolves an active CPO from its current app ID, validates the proposed
        profile and password, stores only an Argon2id password hash, and queues
        an encrypted email-OTP job. No identity or customer exists yet.
      parameters:
        - $ref: "#/components/parameters/CPOAppIDHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerSignupRequest"
      responses:
        "202":
          description: Durable signup challenge and mail job created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeResponse"
        "400":
          description: Invalid JSON, app-ID header, email, password, name, or phone.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: "`signup_unavailable` — app ID does not resolve an active CPO."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/MailUnavailable"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/app/auth/signup/verify:
    post:
      tags: [App User Authentication]
      operationId: verifyCustomerSignup
      summary: Verify email and create the CPO customer
      description: |
        Verifies and consumes the CPO-bound OTP. In one transaction it creates
        a global identity or safely reuses an active one, creates the
        CPO-scoped customer and zero-balance INR wallet, and records an audit
        event. Existing identity credentials and profile are never overwritten.
        This endpoint does not issue customer login tokens.
      parameters:
        - $ref: "#/components/parameters/CPOAppIDHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChallengeVerificationRequest"
      responses:
        "201":
          description: Customer and wallet created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerSignupResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/InvalidChallenge"
        "403":
          description: "`signup_unavailable` — CPO or existing identity is inactive."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: "`customer_already_registered` — verified identity is already this CPO's customer."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/app/auth/signup/resend:
    post:
      tags: [App User Authentication]
      operationId: resendCustomerSignupOTP
      summary: Replace an eligible signup challenge
      description: |
        After the cooldown, invalidates an unconsumed challenge and creates a
        replacement carrying the same password hash and proposed profile. CPO
        app identity and active lifecycle are revalidated.
      parameters:
        - $ref: "#/components/parameters/CPOAppIDHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChallengeResendRequest"
      responses:
        "202":
          description: Replacement challenge and mail job created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/InvalidChallenge"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/MailUnavailable"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/app/auth/login:
    post:
      tags: [App User Authentication]
      operationId: startCustomerLogin
      summary: Validate a customer password and start mail OTP
      description: |
        Resolves the active CPO by app ID, validates the global password and an
        active customer relationship, then durably queues a customer login OTP.
        Invalid identity, password, customer membership, lockout, and customer
        status collapse to `invalid_credentials`.
      parameters:
        - $ref: "#/components/parameters/CPOAppIDHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerLoginRequest"
      responses:
        "202":
          description: Customer login challenge and encrypted mail job created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          description: "`invalid_credentials` — credentials or customer authority are invalid."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: "`signup_unavailable` — app ID does not resolve an active CPO."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/MailUnavailable"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/app/auth/login/verify:
    post:
      tags: [App User Authentication]
      operationId: verifyCustomerLogin
      summary: Verify customer login OTP and create a session
      parameters:
        - $ref: "#/components/parameters/CPOAppIDHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChallengeVerificationRequest"
      responses:
        "200":
          description: Customer session and rotating refresh lineage created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerTokenResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/InvalidChallenge"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/app/auth/login/resend:
    post:
      tags: [App User Authentication]
      operationId: resendCustomerLoginOTP
      summary: Replace an eligible customer login challenge
      parameters:
        - $ref: "#/components/parameters/CPOAppIDHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChallengeResendRequest"
      responses:
        "202":
          description: Replacement customer login challenge created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/InvalidChallenge"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/MailUnavailable"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/app/auth/refresh:
    post:
      tags: [App User Authentication]
      operationId: rotateCustomerRefreshToken
      summary: Rotate a customer refresh token
      description: |
        Accepts the current opaque token once, revalidates user, customer, CPO,
        session, and app-ID state, and atomically returns a replacement.
        Consumed-token reuse revokes the complete customer session.
      parameters:
        - $ref: "#/components/parameters/CPOAppIDHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Customer refresh token rotated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerTokenResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          description: "`invalid_refresh_token` — invalid, expired, revoked, reused, or no longer authorized."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/app/auth/password/forgot:
    post:
      tags: [App User Authentication]
      operationId: startCustomerPasswordRecovery
      summary: Start enumeration-safe customer password recovery
      description: |
        Creates the challenge and encrypted mail job for an eligible customer.
        The generic response contains no challenge material. The eligible
        recipient's email contains the recovery ID, code, and expiry required
        by reset-resend or completion.
      parameters:
        - $ref: "#/components/parameters/CPOAppIDHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "202":
          description: Generic response whether or not an eligible customer exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "403":
          description: App ID does not resolve an active CPO.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/MailUnavailable"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/app/auth/password/reset/resend:
    post:
      tags: [App User Authentication]
      operationId: resendCustomerPasswordResetOTP
      summary: Replace an eligible customer password-reset challenge
      description: |
        Requires the recovery ID delivered to the eligible recipient by email.
        Success invalidates the old pair and returns a replacement challenge;
        the replacement email carries the new ID and code.
      parameters:
        - $ref: "#/components/parameters/CPOAppIDHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChallengeResendRequest"
      responses:
        "202":
          description: Replacement password-reset challenge created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/InvalidChallenge"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/MailUnavailable"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/app/auth/password/reset:
    post:
      tags: [App User Authentication]
      operationId: resetCustomerPassword
      summary: Consume a customer reset OTP and replace the global password
      description: |
        Uses the recovery ID and code delivered to the eligible recipient by
        email. Success revokes every session for the global identity.
      parameters:
        - $ref: "#/components/parameters/CPOAppIDHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password replaced and every identity session revoked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid request or password policy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          $ref: "#/components/responses/InvalidChallenge"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/app/auth/me:
    get:
      tags: [App User Authentication]
      operationId: getCurrentCustomer
      summary: Get the current app-user bootstrap payload
      description: |
        Returns the validated global identity, CPO customer relationship, CPO
        application identity, and customer wallet. Backend handlers obtain the
        same trusted IDs through the `customerauth.Current*` helpers.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CPOAppIDHeader"
      responses:
        "200":
          description: Current customer application context.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerMeResponse"
        "400":
          $ref: "#/components/responses/MissingCPOAppID"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: "`cpo_app_id_mismatch` or wrong session plane."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/app/auth/sessions:
    get:
      tags: [App User Authentication]
      operationId: listCustomerSessions
      summary: List active sessions for this CPO customer
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CPOAppIDHeader"
      responses:
        "200":
          description: Only sessions for the current customer and CPO.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerSessionListResponse"
        "400":
          $ref: "#/components/responses/MissingCPOAppID"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/CPOIntegrationForbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/app/auth/sessions/{session_id}:
    delete:
      tags: [App User Authentication]
      operationId: revokeCustomerSession
      summary: Revoke an owned session for this CPO customer
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CPOAppIDHeader"
        - $ref: "#/components/parameters/SessionID"
      responses:
        "204":
          description: Session and unused refresh token revoked.
        "400":
          description: Missing app ID or invalid session UUID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/CPOIntegrationForbidden"
        "404":
          description: Session absent or belongs to another customer/CPO.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/app/auth/logout:
    post:
      tags: [App User Authentication]
      operationId: logoutCustomerSession
      summary: Revoke the current customer session
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CPOAppIDHeader"
      responses:
        "204":
          description: Current customer session revoked.
        "400":
          $ref: "#/components/responses/MissingCPOAppID"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/CPOIntegrationForbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/app/auth/logout-all:
    post:
      tags: [App User Authentication]
      operationId: logoutAllCustomerSessions
      summary: Revoke every session for this customer in this CPO
      description: Administrative and other-CPO customer sessions are not revoked.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CPOAppIDHeader"
      responses:
        "204":
          description: All sessions for the current CPO customer revoked.
        "400":
          $ref: "#/components/responses/MissingCPOAppID"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/CPOIntegrationForbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/app/auth/password/change:
    post:
      tags: [App User Authentication]
      operationId: changeCustomerPassword
      summary: Change the global identity password
      description: Success revokes every session for the identity in every plane.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CPOAppIDHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed and every identity session revoked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid JSON, password policy, or reused password.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Missing/invalid session or incorrect current password.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          $ref: "#/components/responses/CPOIntegrationForbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/auth/login:
    post:
      tags: [Authentication]
      operationId: startAdministrativeLogin
      summary: Validate password and start an email-OTP challenge
      description: |
        Starts either a platform or a single-CPO administrative login. Platform
        login must omit `cpo_id`; CPO login must include a non-zero `cpo_id`.
        Invalid email, password, scope, identity state, platform authority,
        membership, CPO status, and lockout all collapse to
        `invalid_credentials` to avoid authority enumeration.

        A successful response means the encrypted OTP mail job and challenge
        were committed. It does not mean the provider has delivered the email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/PlatformLoginRequest"
                - $ref: "#/components/schemas/CPOLoginRequest"
            examples:
              platform:
                value:
                  email: superadmin@example.com
                  password: ExampleOnly-Change-Me
                  scope: PLATFORM
              cpo:
                value:
                  email: admin@example.com
                  password: ExampleOnly-Change-Me
                  scope: CPO
                  cpo_id: c821a013-5041-42f7-80c8-aa153cf9d455
      responses:
        "202":
          description: OTP challenge created and mail queued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          description: "`invalid_credentials` — credentials or requested authority are invalid."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/MailUnavailable"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/auth/refresh:
    post:
      tags: [Authentication]
      operationId: rotateRefreshToken
      summary: Rotate a refresh token and issue a new access token
      description: |
        Accepts the current opaque refresh token once. Success stores a
        replacement, marks the submitted token used, updates safe client
        metadata, revalidates identity/session/authority/CPO state, and returns
        the replacement. Reuse of a consumed token revokes the whole session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Refresh token rotated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          description: "`invalid_refresh_token` — token invalid, expired, revoked, reused, or authority changed."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/auth/me:
    get:
      tags: [Authentication]
      operationId: getCurrentPrincipal
      summary: Get the current identity and selected authority
      description: |
        Returns global user metadata and the authority selected by this session.
        CPO sessions include CPO ID, role, current app ID, and app-ID mode;
        platform sessions omit those tenant fields.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current validated principal.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/auth/sessions:
    get:
      tags: [Authentication]
      operationId: listSessions
      summary: List the identity's active unexpired sessions
      description: |
        Returns active platform and CPO sessions for the current global
        identity, newest first. `is_current` identifies the bearer session.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Active sessions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWE
      description: Short-lived signed-then-encrypted administrative access token.
    CPOAppID:
      type: apiKey
      in: header
      name: X-CPO-App-ID
      description: |
        Current app ID for the CPO already established by BearerAuth. This is
        routing metadata, not an authentication secret or tenant selector.
...